+++ app/Http/Controllers/ProductController.php
@@ -5,67 +5,78 @@ namespace App\Http\Controllers;
 use App\Models\Product;
 use App\Models\Category;
 use App\Models\Subcategory;
+use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Yajra\DataTables\Facades\DataTables;
-use Illuminate\Support\Str;
 use Illuminate\Support\Facades\Validator;
 
 class ProductController extends Controller
 {
     public function showProduct()
     {
-        $product = Product::Product();
-        return response()->json(['status' => true, 'product' => $product]);
-        //  view('admin.product.product', compact('product'));
+        if (Gate::allows('is_admin')) {
+
+            $product = Product::Product();
+            return response()->json(['status' => true, 'product' => $product]);
+            //  view('admin.product.product', compact('product'));
+        } else {
+            return view('admin.dashboard');
+        }
     }
     // //////////////////////
 
     public function addProduct(Request $request)
     {
-        $validator = Validator::make($request->all(), [
-            "idcategory" => "required|exists:categories,id",
-            "idsubcategory" => "required|exists:subcategories,id",
-            "productname" => "required|string",
-            "productimage" => "required|image|mimes:jpeg,png,jpg|max:2048",
-            "price" => "required|integer",
-            "description" => "required|string",
-        ]);
+        if (Gate::allows('is_admin')) {
 
-        if ($validator->fails()) {
+            $validator = Validator::make($request->all(), [
+                "idcategory" => "required|exists:categories,id",
+                "idsubcategory" => "required|exists:subcategories,id",
+                "productname" => "required|string",
+                "productimage" => "required|image|mimes:jpeg,png,jpg|max:2048",
+                "price" => "required|integer",
+                "description" => "required|string",
+            ]);
+
+            if ($validator->fails()) {
+                return response()->json([
+                    "status" => false,
+                    "error" => $validator->errors()
+                ], 422);
+            }
+
+            $validated = $validator->validated();
+
+
+            if ($request->hasFile('productimage')) {
+                $image = $request->file('productimage');
+                $imageName = time() . '.' . $image->getClientOriginalExtension();
+                $image->move(public_path('images/'), $imageName);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'error' => 'Product image is required.'
+                ], 422);
+            }
+            $product = Product::addProduct(
+                $validated['idcategory'],
+                $validated['idsubcategory'],
+                $validated['productname'],
+                $imageName,
+                $validated['description'],
+                $validated['price']
+            );
+
             return response()->json([
-                "status" => false,
-                "error" => $validator->errors()
-            ], 422);
-        }
-
-        $validated = $validator->validated();
-
-
-        if ($request->hasFile('productimage')) {
-            $image = $request->file('productimage');
-            $imageName = time() . '.' . $image->getClientOriginalExtension();
-            $image->move(public_path('images/'), $imageName);
+                'status' => true,
+                'product' => $product
+            ]);
         } else {
-            return response()->json([
-                'status' => false,
-                'error' => 'Product image is required.'
-            ], 422);
+            return view('admin.dashboard');
         }
-        $product = Product::addProduct(
-            $validated['idcategory'],
-            $validated['idsubcategory'],
-            $validated['productname'],
-            $imageName,
-            $validated['description'],
-            $validated['price']
-        );
-
-        return response()->json([
-            'status' => true,
-            'product' => $product
-        ]);
     }
 
 
@@ -89,44 +100,60 @@ class ProductController extends Controller
     // //////////////////////////update product
     public function updateProduct(Request $request, $id)
     {
-        $data = $request->only(['idcategory', 'idsubcategory', 'productname', 'price', 'description', 'imageName']);
-        if ($request->hasFile('productimage')) {
-            $image = $request->file('productimage');
-            $imageName = time() . '.' . $image->getClientOriginalExtension();
-            $image->move(public_path('images/'), $imageName);
-        } else {
+        if (Gate::allows('is_admin')) {
+
+            $data = $request->only(['idcategory', 'idsubcategory', 'productname', 'price', 'description', 'imageName']);
+            if ($request->hasFile('productimage')) {
+                $image = $request->file('productimage');
+                $imageName = time() . '.' . $image->getClientOriginalExtension();
+                $image->move(public_path('images/'), $imageName);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'error' => 'Product image is required.'
+                ], 422);
+            }
+            $product = Product::updateProduct($id, $data);
             return response()->json([
-                'status' => false,
-                'error' => 'Product image is required.'
-            ], 422);
+                'status' => true,
+                'product' => $product
+            ]);
+        } else {
+            return view('admin.dashboard');
         }
-        $product = Product::updateProduct($id, $data);
-        return response()->json([
-            'status' => true,
-            'product' => $product
-        ]);
     }
 
     // ///////////////////////
     public function productDelete($id)
     {
-        $product = Product::Deletes($id);
-        // dd($product);
-        $product->delete();
-        return response()->json([
-            'status' => true,
-            'category' => $product
-        ]);
+        if (Gate::allows('is_admin')) {
+
+            $product = Product::Deletes($id);
+            // dd($product);
+            $product->delete();
+            return response()->json([
+                'status' => true,
+                'category' => $product
+            ]);
+        } else {
+            return view('admin.dashboard');
+        }
     }
 
 
     // ////////////////////////
     public function Show()
     {
-        return view('admin.Product.product');
+        if (Gate::allows('is_admin')) {
+
+            return view('admin.Product.product');
+        } else {
+            return view('admin.dashboard');
+        }
     }
     public function getSubcategories($category_id)
     {
+
         $subcategories = Subcategory::where('idcategory', $category_id)->get();
         return response()->json($subcategories);
     }
@@ -134,29 +161,34 @@ class ProductController extends Controller
 
     public function Product(Request $request)
     {
-        $product = new Product();
-        $product->productname = $request->productname;
-        $product->price = $request->price;
-        $product->description = $request->description;
-        $product->idcategory = $request->idcategory;
-        $product->idsubcategory = $request->idsubcategory;
-        if ($request->hasFile('productimage')) {
-            $image = $request->file('productimage');
-            $filename = time() . '.' . $image->getClientOriginalExtension();
-            $image->move(public_path('images'), $filename);
-            $product->productimage = $filename;
-        }
+        if (Gate::allows('is_admin')) {
 
-        $product->save();
-        if (isset($filename)) {
-            $product->image()->create([
-                'url' => 'images/' . $filename
-            ]);
-        }
-        if ($product) {
-            return redirect()->route('product.view')->with('success', 'Product successfully inserted');
+            $product = new Product();
+            $product->productname = $request->productname;
+            $product->price = $request->price;
+            $product->description = $request->description;
+            $product->idcategory = $request->idcategory;
+            $product->idsubcategory = $request->idsubcategory;
+            if ($request->hasFile('productimage')) {
+                $image = $request->file('productimage');
+                $filename = time() . '.' . $image->getClientOriginalExtension();
+                $image->move(public_path('images'), $filename);
+                $product->productimage = $filename;
+            }
+
+            $product->save();
+            if (isset($filename)) {
+                $product->image()->create([
+                    'url' => 'images/' . $filename
+                ]);
+            }
+            if ($product) {
+                return redirect()->route('product.view')->with('success', 'Product successfully inserted');
+            } else {
+                return redirect()->back()->with('error', 'Product has not been inserted');
+            }
         } else {
-            return redirect()->back()->with('error', 'Product has not been inserted');
+            return view('admin.dashboard');
         }
     }
 
@@ -235,10 +267,16 @@ class ProductController extends Controller
 
     public function Editproduct(string $id)
     {
-        $product = Product::find($id);
-        $category = Category::all();
-        $subcategory = Subcategory::where('idcategory', $product->idcategory)->get();
-        return view('admin.Product.editproduct', compact('product', 'category', 'subcategory'));
+        if (Gate::allows('is_admin')) {
+
+
+            $product = Product::find($id);
+            $category = Category::all();
+            $subcategory = Subcategory::where('idcategory', $product->idcategory)->get();
+            return view('admin.Product.editproduct', compact('product', 'category', 'subcategory'));
+        } else {
+            return view('admin.dashboard');
+        }
     }
 
 
@@ -285,36 +323,41 @@ class ProductController extends Controller
 
     public function ProductUpdate(Request $request)
     {
-        // dd($request->all());
-        $product = Product::findOrFail($request->id);
-        // If the product doesn't exist, return an error
-        if (!$product) {
-            return redirect()->back()->with('error', 'Product not found');
+        if (Gate::allows('is_admin')) {
+
+            // dd($request->all());
+            $product = Product::findOrFail($request->id);
+            // If the product doesn't exist, return an error
+            if (!$product) {
+                return redirect()->back()->with('error', 'Product not found');
+            }
+
+
+            $product->idcategory = $request->idcategory;
+            $product->idsubcategory = $request->idsubcategory;
+            $product->productname = $request->productname;
+            $product->price = $request->productprice;
+            $product->description = $request->description;
+
+            // if ($request->hasFile('productimage')) {
+            //     // Delete old image if exists
+            //     $old_image_path = public_path('images/' . $product->productimage);
+            //     if (file_exists($old_image_path)) {
+            //         unlink($old_image_path);
+            //     }
+
+
+            $image = $request->file('productimage');
+            $filename = time() . '.' . $image->getClientOriginalExtension();
+            $image->move(public_path('images'), $filename);
+            $product->productimage = $filename;
+            $product->save();
+
+            // Return a success message
+            return redirect()->route('product.view')->with('success', 'Product successfully updated');
+        } else {
+            return view('admin.dashboard');
         }
-
-
-        $product->idcategory = $request->idcategory;
-        $product->idsubcategory = $request->idsubcategory;
-        $product->productname = $request->productname;
-        $product->price = $request->productprice;
-        $product->description = $request->description;
-
-        // if ($request->hasFile('productimage')) {
-        //     // Delete old image if exists
-        //     $old_image_path = public_path('images/' . $product->productimage);
-        //     if (file_exists($old_image_path)) {
-        //         unlink($old_image_path);
-        //     }
-
-
-        $image = $request->file('productimage');
-        $filename = time() . '.' . $image->getClientOriginalExtension();
-        $image->move(public_path('images'), $filename);
-        $product->productimage = $filename;
-        $product->save();
-
-        // Return a success message
-        return redirect()->route('product.view')->with('success', 'Product successfully updated');
     }
 
     // Delete Product ////////////////////
@@ -343,18 +386,28 @@ class ProductController extends Controller
 
     public function Delete($id)
     {
-        $product = Product::findOrFail($id);
-        $product->delete();
+        if (Gate::allows('is_admin')) {
 
-        return response()->json([
-            'status' => 'success',
-            'message' => 'Product deleted successfully.'
-        ]);
+            $product = Product::findOrFail($id);
+            $product->delete();
+
+            return response()->json([
+                'status' => 'success',
+                'message' => 'Product deleted successfully.'
+            ]);
+        } else {
+            return view('admin.dashboard');
+        }
     }
     public function Multishow(string $id)
     {
-        $multi = Product::with('image')->find($id);
-        $type = 'product';
-        return view('admin.Multipleview.pro', compact('multi', 'type'));
+        if (Gate::allows('is_admin')) {
+
+            $multi = Product::with('image')->find($id);
+            $type = 'product';
+            return view('admin.Multipleview.pro', compact('multi', 'type'));
+        } else {
+            return view('admin.dashboard');
+        }
     }
 }
